[V10.1.10] Use float instead of np.float32 for ExplainableQuantity values to avoid decimal issues (like 1.2 turning into 1.1999999999999997).

+++ efootprint/core/hardware/server_base.py
@@ -243,7 +243,7 @@ class ServerBase(InfraHardware):
                         left_parent=self.raw_nb_of_instances, right_parent=self.fixed_nb_of_instances)
             else:
                 nb_of_instances_np = Quantity(
-                    max_nb_of_instances.magnitude * np.ones(len(self.raw_nb_of_instances), dtype=np.float32),
+                    np.float32(max_nb_of_instances.magnitude) * np.ones(len(self.raw_nb_of_instances), dtype=np.float32),
                     u.dimensionless)
 
                 nb_of_instances = ExplainableHourlyQuantities(
